Added cast.
git-svn-id: svn://svn.code.sf.net/p/peano/code/trunk/src@1268 aa977e5b-e928-4c84-aa66-c90e40304b12

+++ tarch/logging/LogFilterFileReader.cpp
@@ -50,7 +50,7 @@ bool tarch::logging::LogFilterFileReader::parseLine(std::ifstream& file, const s
   const int NumberOfTokensPerLine = 4;
   std::string tokens[NumberOfTokensPerLine];
   int currentToken = 0;
-  while (currentToken<NumberOfTokensPerLine && !file.eof() && characterPosition < lineWithNewline.length()) {
+  while (currentToken<NumberOfTokensPerLine && !file.eof() && characterPosition < static_cast<int>(lineWithNewline.length())) {
     char buffer = lineWithNewline[characterPosition];
 
     const bool bufferIsWhiteSpace =
